Navigation and pages started
Add form started, but still need to incorporate Issuer and recipient info on there to be able to make the button work. There is an error right now due to the fact that i am not using view model yet since creating an object invoice isn't going to work because we need Issuer and recipient info.